move core logic into client class
